+++ authenticator.go
@@ -1,31 +1,32 @@
 package sink
 
 import (
-	"github.com/streamingfast/substreams/client"
 	"os"
+
+	"github.com/streamingfast/substreams/client"
 )
 
-type subsAuthenticator struct {
+type authenticator struct {
 	apiKeyEnvVar   string
 	apiTokenEnvVar string
 }
 
-func NewSubsAuthenticator(apiKeyEnvVar string, apiTokenEnvVar string) *subsAuthenticator {
-	return &subsAuthenticator{
+func newAuthenticator(apiKeyEnvVar string, apiTokenEnvVar string) *authenticator {
+	return &authenticator{
 		apiKeyEnvVar:   apiKeyEnvVar,
 		apiTokenEnvVar: apiTokenEnvVar,
 	}
 }
 
-func (a *subsAuthenticator) GetApiKey() string {
+func (a *authenticator) GetApiKey() string {
 	return a.apiKeyEnvVar
 }
 
-func (a *subsAuthenticator) GetApiToken() string {
+func (a *authenticator) GetApiToken() string {
 	return a.apiTokenEnvVar
 }
 
-func (a *subsAuthenticator) GetAuth() (authToken string, authType client.AuthType) {
+func (a *authenticator) GetTokenAndType() (authToken string, authType client.AuthType) {
 	apiKeyFromEnv := os.Getenv(a.apiKeyEnvVar)
 	if apiKeyFromEnv != "" {
 		return apiKeyFromEnv, client.ApiKey